# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema POS SaaS Multi-Tenant",
    "version": "5.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "List Customers",
        "operationId": "list_customers_api_v1_customers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerResponse"
                  },
                  "title": "Response List Customers Api V1 Customers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Create Customer",
        "operationId": "create_customer_api_v1_customers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pos/checkout": {
      "post": {
        "tags": [
          "POS Checkout"
        ],
        "summary": "Checkout",
        "operationId": "checkout_api_v1_pos_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/pos/sync": {
      "post": {
        "tags": [
          "POS Checkout"
        ],
        "summary": "Sync Offline Sales",
        "operationId": "sync_offline_sales_api_v1_pos_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SaleResponse"
                  },
                  "type": "array",
                  "title": "Response Sync Offline Sales Api V1 Pos Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cash/current": {
      "get": {
        "tags": [
          "Cash Register"
        ],
        "summary": "Get Current Cash Register",
        "operationId": "get_current_cash_register_api_v1_cash_current_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CashRegisterResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Current Cash Register Api V1 Cash Current Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cash/open": {
      "post": {
        "tags": [
          "Cash Register"
        ],
        "summary": "Open Cash Register",
        "operationId": "open_cash_register_api_v1_cash_open_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenCashRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashRegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cash/close": {
      "post": {
        "tags": [
          "Cash Register"
        ],
        "summary": "Close Cash Register",
        "operationId": "close_cash_register_api_v1_cash_close_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseCashRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashRegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cash/movement": {
      "post": {
        "tags": [
          "Cash Register"
        ],
        "summary": "Add Cash Movement",
        "operationId": "add_cash_movement_api_v1_cash_movement_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashMovementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CashMovementCreate": {
        "properties": {
          "cash_register_id": {
            "type": "string",
            "title": "Cash Register Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "cash_register_id",
          "type",
          "amount",
          "reason"
        ],
        "title": "CashMovementCreate"
      },
      "CashRegisterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "opening_balance": {
            "type": "number",
            "title": "Opening Balance"
          },
          "closing_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing Balance"
          },
          "expected_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Balance"
          },
          "difference": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difference"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "user_id",
          "name",
          "status",
          "opening_balance",
          "closing_balance",
          "expected_balance",
          "difference"
        ],
        "title": "CashRegisterResponse"
      },
      "CheckoutRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SaleItemCreate"
            },
            "type": "array",
            "title": "Items"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method",
            "default": "cash"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "cash_register_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Register Id"
          },
          "discount": {
            "type": "number",
            "title": "Discount",
            "default": 0.0
          },
          "offline_sale_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offline Sale Id"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CheckoutRequest"
      },
      "CloseCashRegisterRequest": {
        "properties": {
          "closing_balance": {
            "type": "number",
            "title": "Closing Balance"
          }
        },
        "type": "object",
        "required": [
          "closing_balance"
        ],
        "title": "CloseCashRegisterRequest"
      },
      "CustomerCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tax_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CustomerCreate"
      },
      "CustomerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tax_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "name",
          "tax_id",
          "email",
          "phone",
          "points"
        ],
        "title": "CustomerResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "OpenCashRegisterRequest": {
        "properties": {
          "opening_balance": {
            "type": "number",
            "title": "Opening Balance"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": "Caja Principal"
          }
        },
        "type": "object",
        "required": [
          "opening_balance"
        ],
        "title": "OpenCashRegisterRequest"
      },
...
      "SaleItemCreate": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "unit_price": {
            "type": "number",
            "title": "Unit Price"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "product_name",
          "quantity",
          "unit_price"
        ],
        "title": "SaleItemCreate"
      },
      "SaleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "invoice_number": {
            "type": "string",
            "title": "Invoice Number"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal"
          },
          "discount": {
            "type": "number",
            "title": "Discount"
          },
          "tax": {
            "type": "number",
            "title": "Tax"
          },
          "total": {
            "type": "number",
            "title": "Total"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "invoice_number",
          "subtotal",
          "discount",
          "tax",
          "total",
          "payment_method",
          "status"
        ],
        "title": "SaleResponse"
      },
      "SyncBatchRequest": {
        "properties": {
          "sales": {
            "items": {
              "$ref": "#/components/schemas/CheckoutRequest"
            },
            "type": "array",
            "title": "Sales"
          }
        },
        "type": "object",
        "required": [
          "sales"
        ],
        "title": "SyncBatchRequest"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user_id",
          "company_id",
          "full_name",
          "role",
          "permissions"
        ],
        "title": "TokenResponse"
      },
      "UserProfile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "email",
          "full_name",
          "role",
          "permissions"
        ],
        "title": "UserProfile"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}